backend/requirements.txt
Flask
flask-cors
pymongo
python-jose
requests
